Address review: DownloadSearchUIModel, country lookup in WorldRegion
- rename and move the search logic to net.osmand.plus.download.ui.DownloadSearchUIModel
- drop the duplicated country lookup, WorldRegion.getCountryRegion() now also resolves
  countries that are placed next to the continents instead of under one
- one result list instead of matched/found, simpler isMatch(), 120 column formatting
- name the report class after the screen it renders
- add London, New york, New and San Fr search examples
