+++ OsmAnd/test/java/net/osmand/plus/download/DownloadSearchScreenReport.java
@@ -2,8 +2,8 @@ package net.osmand.plus.download;
 
 import androidx.annotation.NonNull;
 
-import net.osmand.plus.download.DownloadSearchHelperTest.ReportRow;
-import net.osmand.plus.download.DownloadSearchHelperTest.ReportScreen;
+import net.osmand.plus.download.DownloadSearchUIModelTest.ReportRow;
+import net.osmand.plus.download.DownloadSearchUIModelTest.ReportScreen;
 import net.osmand.util.Algorithms;
 
 import java.text.SimpleDateFormat;
@@ -15,10 +15,10 @@ import java.util.Locale;
 import java.util.Map;
 
 /**
- * Renders the search results collected by {@link DownloadSearchHelperTest} as a standalone page,
+ * Renders the search results collected by {@link DownloadSearchUIModelTest} as a standalone page,
  * so the "Maps & Resources" search screens can be reviewed without taking screenshots.
  */
-class SearchReport {
+class DownloadSearchScreenReport {
 
 	@NonNull
 	static String build(@NonNull List<ReportScreen> screens) {
@@ -35,7 +35,7 @@ class SearchReport {
 				.append("<style>\n").append(css()).append("</style>\n");
 		html.append("<header class=\"page\"><h1>Maps &amp; Resources search results</h1>")
 				.append("<p>Rows produced by the search screen for every query below. Generated by ")
-				.append("DownloadSearchHelperTest on ").append(date).append(".</p></header>\n");
+				.append("DownloadSearchUIModelTest on ").append(date).append(".</p></header>\n");
 
 		for (Map.Entry<String, List<ReportScreen>> entry : byCatalog.entrySet()) {
 			html.append("<section>\n<h2>").append(escape(entry.getKey())).append("</h2>\n")
